move user's libproto to pkg
