+++ user/cmd/main.go
@@ -9,9 +9,9 @@ import (
 	"github.com/GrabItYourself/giys-backend/lib/logger"
 	"github.com/GrabItYourself/giys-backend/lib/postgres"
 	"github.com/GrabItYourself/giys-backend/user/internal/config"
-	"github.com/GrabItYourself/giys-backend/user/internal/libproto"
 	"github.com/GrabItYourself/giys-backend/user/internal/repository"
 	"github.com/GrabItYourself/giys-backend/user/internal/server"
+	"github.com/GrabItYourself/giys-backend/user/pkg/libproto"
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/reflection"
