+++ user/pkg/client/client.go
@@ -1,12 +1,12 @@
 package client
 
 import (
-	"github.com/GrabItYourself/giys-backend/user/internal/libproto"
+	"github.com/GrabItYourself/giys-backend/user/pkg/libproto"
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
 )
 
-func NewClient(addr string, opts ...grpc.DialOption) (*libproto.UserServiceClient, error) {
+func NewClient(addr string, opts ...grpc.DialOption) (libproto.UserServiceClient, error) {
 	conn, err := grpc.Dial(addr, opts...)
 	if err != nil {
 		return nil, errors.Wrap(err, "Failed to connect to user service")
@@ -14,5 +14,5 @@ func NewClient(addr string, opts ...grpc.DialOption) (*libproto.UserServiceClien
 	defer conn.Close()
 
 	client := libproto.NewUserServiceClient(conn)
-	return &client, nil
+	return client, nil
 }
